Moved projectList, Refactored ProjectCard to stay updated with props
