+++ src/views/Projects.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Stack } from "office-ui-fabric-react";
 import { ProjectList as ProjectListElement } from "../components/ProjectList";
-import ProjectList from '../common/projectList.json';
+import ProjectList from '../assets/projectList.json';
 
 export const Projects: React.StatelessComponent = () => {
   return (
